% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{accuracy_stats}
\alias{accuracy_stats}
\title{Compute common accuracy and agreement metrics}
\usage{
accuracy_stats(.data, true_values, ..., include_counts = FALSE)
}
\description{
Given a vector of \code{true_values} and one or more vectors of test values
(passed to \code{...}), computes sensitivity, specificity, positive predictive
value (PPV), negative predictive value (NPV), and Cohen's kappa.
}
\examples{
# create example data w predictors with different properties:
ex_data <- tibble(
  actual = rbinom(250, 1, .3),                                   # 250 cases, p(outcome) = .3
  prediction1 = if_else(runif(250) <= .05, 1L - actual, actual), # 5\% error rate
  prediction2 = if_else(runif(250) <= .15, 1L - actual, actual), # 15\% error rate
  prediction3 = if_else(runif(250) <= .35, 1L - actual, actual), # 35\% error rate
  prediction4 = if_else(runif(250) <= .15, 1L, actual),          # 15\% with positive bias
  prediction5 = if_else(runif(250) <= .15, 0L, actual)           # 15\% with negative bias
)

# testing predicted v actual values
ex_data \%>\%
  accuracy_stats(actual, prediction1)

# can test multiple predictors simultaneously
ex_data \%>\%
  accuracy_stats(actual, prediction1:prediction5)

# if `include_counts` = TRUE, will also return n of false positives,
# false negatives, etc., as well as and observed and expected \% agreement
ex_data \%>\%
  accuracy_stats(actual, prediction1:prediction5, include_counts = TRUE)

# cleanup
rm(ex_data)

}
