% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{add_crossings}
\alias{add_crossings}
\title{Add crossings to a dataframe for area charts}
\usage{
add_crossings(data, x, y, .by = NULL)
}
\arguments{
\item{data}{A data frame containing the original x and y values.}

\item{x}{x-axis values.}

\item{y}{y-axis values.}

\item{.by}{Grouping variable(s). Useful for computing crossings for faceted plots.}
}
\value{
The input data frame with additional rows representing the crossings (where y = 0), and two new columns:
\itemize{
\item \code{pos_neg}: Indicates whether the y-value is positive ("pos") or negative ("neg").
\item \code{cross_grp}: A grouping variable for the segments between crossings.
}
}
\description{
Augments a dataframe with x-values where y = f(x) = 0. This is useful for creating area charts with different fills for values less than versus greater than 0.
}
\details{
The returned dataframe will include columns \code{pos_neg} and \code{cross_group}. Within \code{geom_area()}, \code{cross_group} should be mapped to \code{group}, while \code{pos_neg} should be mapped to aesthetics such as \code{fill} or \code{color}.
}
\examples{
nile_flow <- tibble::tibble(
  Year = time(Nile),
  Flow = as.numeric(Nile),
  FlowDelta = (Flow - Flow[[1]]) / Flow[[1]]
)
nile_flow_x0 <- add_crossings(nile_flow, Year, FlowDelta)
ggplot2::ggplot(nile_flow_x0, ggplot2::aes(Year, FlowDelta)) +
  ggplot2::geom_area(
    ggplot2::aes(group = cross_grp, color = pos_neg, fill = pos_neg),
    alpha = 0.25,
    show.legend = FALSE
  ) +
  ggplot2::geom_hline(yintercept = 0, linewidth = 0.25) +
  ggplot2::scale_color_manual(
    values = c("darkred", "blue"),
    aesthetics = c("color", "fill")
  ) +
  ggplot2::scale_y_continuous(
    "Nile River Annual Flow:\n\% Change from 1871",
    labels = scales::percent
  ) +
  ggplot2::theme_minimal()

}
