% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_header}
\alias{add_header}
\title{Add header rows to a table}
\usage{
add_header(
  data,
  from,
  to,
  skip_single_row = FALSE,
  indent = "",
  drop_from = TRUE
)
}
\description{
Inserts header rows in \code{to} using unique values of \code{from}.
}
\examples{
dplyr::starwars \%>\%
  head(13) \%>\%
  dplyr::arrange(species) \%>\%
  add_header(from = species, to = name, indent = "  ")

}
