% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{add_plot_slide}
\alias{add_plot_slide}
\title{Add a plot to a PowerPoint slide}
\usage{
add_plot_slide(
  pptx,
  title = NULL,
  plot = ggplot2::last_plot(),
  w = 7,
  h = 4,
  bg = "white",
  w_margin = 0.15,
  h_margin = 0.15,
  layout = "Title and Content",
  ...
)
}
\arguments{
\item{pptx}{an object of class rpptx, as created by officer::read_pptx()}

\item{title}{optional slide title}

\item{plot}{the plot to add. Default is the last plot created.}

\item{w}{plot width in inches}

\item{h}{plot height in inches}

\item{bg}{background color of the plot area}

\item{w_margin}{horizontal margin in inches}

\item{h_margin}{vertical margin in inches}

\item{layout}{slide layout to use}

\item{...}{additional arguments passed to officer::ph_with()}
}
\value{
An updated rpptx object
}
\description{
This function adds a new slide to a PowerPoint presentation with a plot
centered beneath the title, scaled to be as large as possible within the
specified margins while preserving aspect ratio.
}
\examples{
\dontrun{
library(ggplot2)
library(officer)

plot <- ggplot(mtcars, aes(mpg, wt)) +
  geom_point()

pptx <- read_pptx()
pptx <- pptx |>
add_plot_slide("Larger Elements", plot, w = 5, h = 3.5) |>
  add_plot_slide("Smaller Elements", plot, w = 10, h = 7) |>
  add_plot_slide("Wider", plot, w = 9, h = 3) |>
  add_plot_slide("Taller", plot, w = 4, h = 6)

path <- paste0(tempfile(), ".pptx")
print(pptx, target = path)
file.open(path)

invisible(file.remove(path))
}

}
