% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{aggregate_if_any}
\alias{aggregate_if_any}
\alias{sum_if_any}
\alias{max_if_any}
\alias{min_if_any}
\title{Sums, maxima and minima with alternative missing value handling}
\usage{
sum_if_any(..., na.rm = TRUE)

max_if_any(..., na.rm = TRUE)

min_if_any(..., na.rm = TRUE)
}
\arguments{
\item{...}{numeric, logical, or (for \code{max_if_any()} and \code{min_if_any()}) character vectors.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\description{
Returns the sum, maximum, or minimum of input values, similar to
\code{base::sum()}, \code{min()}, and \code{max()}. Unlike these base functions, these
variants return \code{NA} when all values are \code{NA} and \code{na.rm = TRUE}.
(\code{base::sum()}, \code{min()}, and \code{max()} return \code{0}, \code{-Inf}, and \code{Inf},
respectively, in this situation). Also unlike base functions, \code{na.rm} is
\code{TRUE} by default (since this is the typical use case).
}
\examples{
some_na <- c(1, 2, NA)
all_na <- c(NA, NA, NA)

# unlike base functions, `na.rm = TRUE` by default
max(some_na)
max_if_any(some_na)

# unlike base functions, returns `NA` when `na.rm = TRUE` and all inputs are `NA`
sum(all_na, na.rm = TRUE)
sum_if_any(all_na)

}
