% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{any-all-in}
\alias{any-all-in}
\alias{\%all_in\%}
\alias{\%any_in\%}
\title{Test whether multiple values are in a vector}
\usage{
lhs \%all_in\% rhs

lhs \%any_in\% rhs
}
\description{
These infix operators test whether any or all left-hand side elements occur
in the right-hand side.
}
\details{
\verb{\%all_in\%} returns \code{TRUE} if all elements of left operand (\code{lhs}) are found
in right operand (\code{rhs}). Equivalent to \code{all(lhs \%in\% rhs)}.

\verb{\%any_in\%} returns \code{TRUE} if any elements of left operand (\code{lhs}) are found
in right operand (\code{rhs}). Equivalent to \code{any(lhs \%in\% rhs)}.
}
\examples{

maybe_states <- c("Idaho", "Illinois", "North Tuba", "Maine")
maybe_states \%any_in\% state.name  # TRUE
maybe_states \%all_in\% state.name  # FALSE
rm(maybe_states)

}
