% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{asterisks}
\alias{asterisks}
\title{Return asterisks corresponding to p-values}
\usage{
asterisks(
  p,
  trends = TRUE,
  levels = c(0.1, 0.05, 0.01, 0.001),
  marks = c(sig = "*", trend = "+", ns = NA_character_),
  include_key = TRUE
)
}
\description{
Return asterisks corresponding to p-values
}
