% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{bizday}
\alias{bizday}
\alias{nth_bizday}
\alias{next_bizday}
\title{Find the nth or next business day}
\usage{
nth_bizday(
  x,
  n,
  include_today = TRUE,
  holidays = c("Chestnut", "Illinois", "federal")
)

next_bizday(
  x,
  include_today = TRUE,
  holidays = c("Chestnut", "Illinois", "federal")
)
}
\arguments{
\item{x}{A date or vector of dates.}

\item{n}{A single integer specifying which future business day to find.}

\item{include_today}{Logical, indicating whether the provided date(s) should be considered as potential dates to return. Defaults to \code{TRUE}.}

\item{holidays}{A character string specifying which set of holidays to use. Defaults to "Chestnut".}
}
\value{
\itemize{
\item \code{nth_bizday} returns the nth business day from the provided date(s).
\item \code{next_bizday} returns the next business day from the provided date(s).
}
}
\description{
These functions find the nth or next business day from a given date, optionally excluding or including the provided date. The set of holidays used to determine business days can be specified.
}
\details{
\code{nth_bizday} returns the nth business day from the given date(s), where n is specified by the \code{n} argument.

\code{next_bizday} is a shortcut for \code{nth_bizday} with \code{n = 1}, returning the next business day from the given date(s).

The \code{holidays} argument specifies the set of holidays to consider when determining business days. Available options are "Chestnut" (default), "Illinois", and "federal".
}
\examples{
nth_bizday(as.Date("2023-07-01"), 5)
next_bizday(as.Date("2023-07-01"))
nth_bizday(as.Date("2023-07-01"), 1, include_today = FALSE)

}
