% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{cascade}
\alias{cascade}
\alias{cascade_fill_bwd}
\alias{cascade_fill_fwd}
\alias{cascade_summarize}
\title{Utilities for service cascades}
\usage{
cascade_fill_bwd(data, vars)

cascade_fill_fwd(data, vars, fill = c("NA", "FALSE"))

cascade_summarize(data, vars)
}
\description{
Functions for working with service cascades.
}
