% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{coerce_na_range}
\alias{coerce_na_range}
\title{Set NA values based on numbers stored as strings.}
\usage{
coerce_na_range(x, ...)
}
\description{
Changes values coercible to numeric in range \code{range_min}:\code{range_max} to \code{NA}.
Useful for imported SPSS files.
}
