% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cohen_w}
\alias{cohen_w}
\title{Compute Cohen's w}
\usage{
cohen_w(chisq)
}
\arguments{
\item{chisq}{An \code{"htest"} object returned by \code{stats::chisq.test()}.}
}
\description{
Cohen's w is an effect size measure for associations between nominal
variables, generally used in conjunction with chi-squared tests. \code{cohen_w()}
computes Cohen's w from results from a chi-squared test.
}
\examples{

chisq_out <- chisq.test(ggplot2::diamonds$cut, ggplot2::diamonds$color)
cohen_w(chisq_out)

}
