% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cols_info}
\alias{cols_info}
\title{Get information about dataframe columns}
\usage{
cols_info(x, zap_spss = TRUE)
}
\description{
Returns a summary of each column's class, type, and missing data. If
dataframe was imported from SPSS .sav and "labelled" package is installed,
SPSS variable labels will also be included.
}
