% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_group.R
\name{count_duplicates}
\alias{count_duplicates}
\title{Count duplicates across specified columns}
\usage{
count_duplicates(.data, ..., na.rm = FALSE)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Columns to use for duplicate checks. If empty, all columns are used.}

\item{na.rm}{If \code{TRUE}, rows containing \code{NA} in the specified columns will be removed before counting duplicates.}
}
\value{
A data frame with columns:
\itemize{
\item \code{instances}: The number of times each unique value is duplicated
\item \code{n_unique}: The number of unique values duplicated \code{instances} times
\item \code{n_total}: The total number of observations duplicated \code{instances} times
}
}
\description{
This variant of \code{\link[dplyr:count]{dplyr::count()}} returns the number of duplicate observations across the specified columns.
Returns the number of unique duplicated values, as well as the total number
of duplicated observations.
}
\examples{
df <- tibble::tibble(
  x = c(1, 1, 2, 3, 3),
  y = c('a', 'a', 'b', 'c', 'c')
)

count_duplicates(df)
count_duplicates(df, x)
count_duplicates(df, y)

}
