% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_group.R
\name{count_na}
\alias{count_na}
\title{Count NA values by group}
\usage{
count_na(
  .data,
  ...,
  .label_missing = NA_character_,
  .label_valid = "OK",
  .add_percent = FALSE
)
}
\description{
Returns patterns of missingness across one or more variables, with number of
cases for each pattern.
}
