% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{count_unique}
\alias{count_unique}
\title{Count number of unique values in a set of dataframe columns}
\usage{
count_unique(.data, ..., name = "n_unique", na.rm = FALSE)
}
\description{
A pipe-friendly wrapper around \code{dplyr::n_distinct()}.
}
