% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{crosstab}
\alias{crosstab}
\title{Cross-tabulate observations}
\usage{
crosstab(.data, ..., .drop = TRUE)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Variables to cross-tabulate. (Only two variables are currently supported.)}

\item{.drop}{If \code{FALSE}, will include counts for empty groups (i.e. for levels
of factors that don't exist in the data).}
}
\description{
Builds a contingency table similar to \code{base::table()}. Unlike
\code{base::table()}, \code{crosstab()} is pipe-friendly, and outputs an ordinary
tibble / data.frame -- e.g., it will retain its structure if exported to csv.

Currently supports only two variables.
}
