% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{datetimes_to_date}
\alias{datetimes_to_date}
\title{Convert all datetimes in table to dates}
\usage{
datetimes_to_date(.data)
}
\description{
Returns a dataframe with all datetime columns (i.e., class \code{POSIXct} or
\code{POSIXlt}) converted to \code{Date}.
}
