% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{df_compare}
\alias{df_compare}
\title{Compare two data frames and show differences}
\usage{
df_compare(x, y, suffix = c(".x", ".y"), keep = NULL)
}
\arguments{
\item{x, y}{a pair of data frames}

\item{suffix}{suffixes to indicate source data frame in output.}

\item{keep}{<\link{tidy-select}> Columns to include in the output even if they have no differences.}
}
\value{
A data frame with rows and columns that have differing values between
\code{x} and \code{y}. Differing columns are included twice, with suffixes appended.
Columns specified in \code{keep} are always included.
}
\description{
Given two data frames with the same dimensions and column order, returns a data frame including only rows and columns with differences.
}
\examples{
x <- data.frame(id = 1:3, A = c(7, 8, 9), B = c(10, 20, 30), C = c("x", "y", "z"))
y <- data.frame(id = 1:3, A = c(7, 8, 99), B = c(10, 20, 30), C = c("X", "y", "Z"))

df_compare(x, y)
df_compare(x, y, keep = id)

}
