% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{discard_na}
\alias{discard_na}
\title{Remove missing values}
\usage{
discard_na(x)
}
\description{
Returns a vector with \code{NA}s removed. Similar to \code{stats::na.omit.default()}, but does not add attributes to returned value.
}
