% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{drop_na_rows}
\alias{drop_na_rows}
\title{Drop rows where all columns are \code{NA}}
\usage{
drop_na_rows(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{(Optional) Columns to test for \code{NA}s. If not specified, all columns will be considered.}
}
\value{
A data frame with rows removed that contain \code{NA} values across all (specified) columns.
}
\description{
Drops rows where all specified columns are \code{NA}. If no columns are specified, all columns will be considered.
}
\examples{
dat <- tibble::tibble(
  x = c(NA, NA, 3),
  y = c(NA, NA, 4),
  z = c(5, NA, NA)
)
drop_na_rows(dat)
drop_na_rows(dat, x, y)

}
