% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{dunn_test}
\alias{dunn_test}
\title{Pairwise post-hoc test following Kruskal-Wallis test}
\usage{
dunn_test(
  x,
  groups,
  data,
  p.adjust.method = c("holm", "hochberg", "bonferroni", "bh", "by", "sidak", "hs",
    "none"),
  alpha = 0.05
)
}
\arguments{
\item{x}{a numeric vector.}

\item{groups}{a vector or factor giving the group for the corresponding elements of \code{x}.}

\item{data}{a data frame containing the variables.}

\item{p.adjust.method}{character. method for adjusting p-values for multiple comparisons.}

\item{alpha}{numeric. alpha level.}
}
\value{
A tibble with columns:
\itemize{
\item \code{contrast}: The compared groups.
\item \code{statistic}: Dunn's test statistic (z).
\item \code{adj.p.value}: Adjusted p-value based on the specified \code{p.adjust.method}.
}
}
\description{
This is a tidy wrapper around \code{\link[dunn.test:dunn.test]{dunn.test::dunn.test()}}. It performs Dunn's test, a non-parametric pairwise follow-up for a Kruskal-Wallis test.
}
\examples{
mtcars2 <- transform(mtcars, cyl = factor(cyl))
kruskal.test(mpg ~ gear, data = mtcars2)
dunn_test(mpg, gear, data = mtcars2)

}
\seealso{
\code{\link[dunn.test:dunn.test]{dunn.test::dunn.test()}}
}
