% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{eq_shape}
\alias{eq_shape}
\title{Test if two objects have the same shape}
\usage{
eq_shape(x, y)
}
\arguments{
\item{x}{An object.}

\item{y}{An object.}
}
\value{
\code{TRUE} if \code{x} and \code{y} have the same shape, \code{FALSE} otherwise.
}
\description{
Checks if two objects \code{x} and \code{y} have the same shape, i.e., the same dimensions or same length for vectors.
}
\examples{
eq_shape(1:5, 1:5)
eq_shape(1:5, 1:6)
eq_shape(matrix(1:6, nrow = 2), matrix(1:6, nrow = 3))
eq_shape(matrix(1:6, nrow = 2), matrix(1:6, nrow = 2))

}
