% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{fct_collapse_alt}
\alias{fct_collapse_alt}
\title{Collapse factor levels with additional controls}
\usage{
fct_collapse_alt(
  .f,
  ...,
  other_level = NULL,
  reorder = TRUE,
  unknown_levels = c("ignore", "warn", "error")
)
}
\arguments{
\item{.f}{A factor (or character vector).}

\item{...}{A series of named character vectors. The levels in each vector
will be replaced with the name.}

\item{other_level}{Value of level used for "other" values not named in \code{...}.
If \code{NULL}, no extra level is created.}

\item{reorder}{If \code{TRUE}, the collapsed levels are ordered in the order they
are listed in \code{...}, followed by \code{other_level} if specified, and then
existing levels.}

\item{unknown_levels}{How to handle levels listed in \code{...} that aren't
present in the input factor \code{.f}. Options are:
\itemize{
\item "ignore": ignore unknown levels without a warning (default),
\item "warn": issue a warning and then ignore unknown levels,
\item "error": raise an error.
}}
}
\value{
A factor with collapsed levels.
}
\description{
Collapses factor levels into manually defined groups like
\code{\link[forcats:fct_collapse]{forcats::fct_collapse()}}, with additional options to control behavior when
specified levels don't exist in the data and to order factor levels in the
order listed.
}
\examples{
f <- factor(c("a", "b", "c", "d", "e", "f"))

fct_collapse_alt(f, EFG = c("e", "f", "g"), AB = c("a", "b"))
fct_collapse_alt(f, EFG = c("e", "f", "g"), AB = c("a", "b"), reorder = FALSE)
fct_collapse_alt(f, EFG = c("e", "f", "g"), AB = c("a", "b"), other_level = "other")
# `unknown_levels = "warn"` mirrors behavior of `forcats::fct_collapse()`
\donttest{
fct_collapse_alt(f, EFG = c("e", "f", "g"), AB = c("a", "b"), unknown_levels = "warn")
}

}
