% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fct_filter}
\alias{fct_filter}
\title{Filter by and drop factor levels simultaneously}
\usage{
fct_filter(.data, .fct, .keep = NULL, .drop = NULL)
}
\description{
Filters dataframe by specified levels of \code{.fct}, then drops filtered levels
from \code{.fct}.
}
