% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{fct_na_if}
\alias{fct_na_if}
\title{Convert specified factor levels to \code{NA}}
\usage{
fct_na_if(x, y)
}
\arguments{
\item{x}{A factor.}

\item{y}{A character vector of levels to convert to \code{NA}.}
}
\value{
The input factor with specified levels converted to \code{NA} and removed from the levels.
}
\description{
Converts specified level(s) of a factor to \code{NA}, removing those levels from
the factor. This differs from the behavior of \code{\link[dplyr:na_if]{dplyr::na_if()}}, which (1)
replaces values with \code{NA} but retains the associated factor level, and (2)
can replace only a single value.
}
\examples{
f <- factor(c("a", "b", "c", "a"))
fct_na_if(f, "a")
fct_na_if(f, c("a", "c"))

# compare `na_if()`
dplyr::na_if(f, "a")

}
\seealso{
{
\code{\link[dplyr:na_if]{dplyr::na_if()}}
}
}
