% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{fct_reorder_n}
\alias{fct_reorder_n}
\title{Reorder factor levels by sorting along multiple other variables.}
\usage{
fct_reorder_n(.f, ..., .desc = FALSE)
}
\description{
Reorders levels of \code{.f} based on variables passed to \code{...}, breaking ties
using each variable in the order passed.
}
