% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{filter_drop}
\alias{filter_drop}
\title{Filter by and drop a column simultaneously}
\usage{
filter_drop(.data, .col, .keep = NULL, .drop = NULL)
}
\description{
Filters dataframe by specified values of \code{.col}, then drops \code{.col}.
}
