% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{find_na_cols}
\alias{find_na_cols}
\alias{drop_na_cols}
\title{Identify or remove columns with no data}
\usage{
find_na_cols(.data, cols = tidyselect::everything())

drop_na_cols(.data, cols = tidyselect::everything(), quietly = FALSE)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a tibble).}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to check.}

\item{quietly}{If \code{FALSE}, will print which columns were dropped.}
}
\description{
\code{find_na_cols()} returns the names of all columns in \code{.data} where all values
are \code{NA}. \code{drop_na_cols()} returns the dataset with all-\code{NA} columns removed.
}
