% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{floor_month}
\alias{floor_month}
\alias{floor_week}
\alias{floor_days}
\title{Floor methods for date objects}
\usage{
floor_month(x)

floor_week(x, week_start = getOption("lubridate.week.start", 7))

floor_days(x, n = 1L, start = min(x, na.rm = TRUE), reset_monthly = FALSE)
}
\description{
\code{floor_month()} and \code{floor_week()} are simple wrappers around
\code{lubridate::floor_date()} that round down to the first day of the month or
week.

\code{floor_days()} rounds down to the nearest \code{n}-day increment. Floors are
defined relative to the earliest date in \code{x}, unless a different start date
is passed to \code{start}. Default behavior differs from
\code{lubridate::floor_date(x, unit = "{n} days")}, which "resets" the floor to
the first of the month with each month. \code{lubridate}-like behavior can be
achieved by setting \code{reset_monthly = TRUE}.
}
