% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gain_missing_codes}
\alias{gain_missing_codes}
\title{Missing codes for GAIN ABS}
\format{
A named numeric vector.
}
\usage{
gain_missing_codes
}
\description{
Labelled missings used in GAIN datasets.
}
\keyword{datasets}
