% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gain_ss_score}
\alias{gain_ss_score}
\title{Compute time period scores for GAIN-SS scales}
\usage{
gain_ss_score(..., .prefix = NULL)
}
\description{
Pass all scale items to \code{...}. Will return columns with the score for
lifetime, past year, past 90 days, and past month positive items
}
