% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_db.R
\name{gen_sql_db}
\alias{gen_sql_db}
\title{gen_sql_db}
\usage{
gen_sql_db(project, mission = "", overw = F)
}
\description{
Function to generate SQL database from the L3 of a project,
locally a SQLite, remotely on a PostgreSQL instalation (optional, not implemented yet)
}
