% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_water_sample_db.R
\name{generate_spm_db}
\alias{generate_spm_db}
\alias{l3_gen_water_sample}
\title{generate_spm_db}
\usage{
l3_gen_water_sample(project = "", mission = "", params = c("SPM", "Ag", "Ap"))
}
\arguments{
\item{project}{The top level folder of the project}

\item{mission}{Optional, name for data base file.
If none is providen, it is taken from the project name.}

\item{params}{Character vector used to select for which parameters data bases should be created}
}
\description{
Generate a data base (one file by parameter) from L1 water sample files.
Different methods are aplicated depending on the type of parameter.
For Ag and Ap parameter, it takes the output of RspectroAbs package.
For SPM (Chl-a to be implemented) it read a .csv such as "link to example table"
}
\author{
Raphael Mabit
}
