% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ggview}
\alias{ggview}
\title{Nicer ggplot rendering}
\usage{
ggview(
  plot = ggplot2::last_plot(),
  width = NULL,
  height = NULL,
  type = c("svg", "png")
)
}
\description{
Saves ggplot object as .svg in R temp directory, then displays in RStudio
Viewer pane. Results in better image quality on Windows machines.
}
