% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{group_split_named}
\alias{group_split_named}
\title{Split dataframe by named groups}
\usage{
group_split_named(
  .data,
  ...,
  .keep = TRUE,
  .sep = ".",
  .col_names = FALSE,
  .col_sep = "_",
  .nested = FALSE,
  .na.rm = FALSE,
  .add_groups = TRUE
)
}
\description{
Divides \code{.data} into a named list of dataframes defined by the grouping
structure. Grouping variables can optionally be passed to \code{...}. A nested
list will be returned if there is more than one grouping variable and
\code{.nested} = \code{TRUE}.
}
\examples{
by_cyl_gear1 <- mtcars \%>\%
  group_split_named(cyl, gear, .col_names = TRUE)
by_cyl_gear1$cyl_6.gear_4

by_cyl_gear2 <- mtcars \%>\%
  group_split_named(cyl, gear, .col_names = TRUE, .nested = TRUE)
by_cyl_gear2$cyl_6
by_cyl_gear2$cyl_6$gear_4

}
