% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holidays_il}
\alias{holidays_il}
\title{Illinois state holidays over a 20-year period}
\format{
A tibble with 252 rows and 2 variables:
\describe{
\item{Date}{date holiday observed}
\item{Holiday}{holiday name}
}
}
\usage{
holidays_il
}
\description{
A dataset containing dates of State of Illinois holidays from 2010-12-31
through 2030-12-31.
}
\keyword{datasets}
