% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holidays_us}
\alias{holidays_us}
\title{Dates of US federal holidays over a 20-year period}
\format{
A tibble with 212 rows and 2 variables:
\describe{
\item{Date}{date holiday observed}
\item{Holiday}{holiday name}
}
}
\source{
\url{https://www.opm.gov/policy-data-oversight/pay-leave/federal-holidays/}
}
\usage{
holidays_us
}
\description{
A dataset containing dates of United States federal holidays from 2010-12-31
through 2030-12-31.
}
\keyword{datasets}
