% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{in_excel}
\alias{in_excel}
\title{Open dataframe in Excel}
\usage{
in_excel(df, name, na = "")
}
\description{
Saves dataframe as .csv in R temp directory, then opens in Excel. The .csv
will have a randomly-generated name unless otherwise specified in \code{name}.
}
