% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{is_TRUE}
\alias{is_TRUE}
\alias{is_FALSE}
\alias{is_TRUE_or_NA}
\alias{is_FALSE_or_NA}
\title{Vectorized logical tests}
\usage{
is_TRUE(x, strict = TRUE)

is_FALSE(x, strict = TRUE)

is_TRUE_or_NA(x, strict = TRUE)

is_FALSE_or_NA(x, strict = TRUE)
}
\arguments{
\item{x}{Vector to be tested}

\item{strict}{If \code{TRUE} (the default), numeric and character types will
always return \code{FALSE}. If \code{FALSE}, numeric and character vectors that can be
coerced to logical (e.g., \code{1}, \code{"FALSE"}) will be coerced before testing.}
}
\description{
\code{is_TRUE()} and \code{is_FALSE()} are vectorized versions of \code{base::isTRUE()} and
\code{base::isFALSE()}, respectively. \code{is_TRUE()} returns \code{TRUE} for each vector
element that evaluates to \code{TRUE}, and \code{FALSE} for all other elements
(including for \code{NA}s and non-logical values). Useful for handling \code{NA}s in
logical tests.
}
