% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_dir_list.R
\name{l2_dir_list}
\alias{l2_dir_list}
\title{Fuction to create directories.for.*.dat requierd by Cops and Riops packages}
\usage{
l2_dir_list(path = getwd(), what = c("COPS", "IOP"), boat = c(""), upInDir = T)
}
\arguments{
\item{path}{is the L2 folder path (string) where uniquely identifie stations can be found.}

\item{what}{is a character vector of what to search for, default c("COPS","IOP")}

\item{boat}{is a character vector of to filter by boat, default c("")}

\item{upInDir}{logical indicating if update of file path append also in L2 subdirectories.}
}
\value{
nothing
}
\description{
Fuction to create directories.for.*.dat requierd by Cops and Riops packages
}
