% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/util.R
\name{median_dbl}
\alias{median_dbl}
\title{Median value as double - DEPRECATED}
\usage{
median_dbl(x, na.rm = FALSE, ...)

median_dbl(x, na.rm = FALSE, ...)
}
\description{
Deprecated in lighthouse v0.7.0. The main use case for this function was to avoid type errors in \code{dplyr::if_else()} and \code{case_when()}, but this is no longer necessary with
\href{https://dplyr.tidyverse.org/news/index.html#vctrs-1-1-0}{changes introduced in dplyr v1.1.0}.

Returns median of \code{x} as a double vector. Alternative to \code{stats::median()}
when a consistent return value is needed.

Returns median of \code{x} as a double vector. Alternative to \code{stats::median()}
when a consistent return value is needed.
}
\examples{
\dontrun{
# stats::median raises error because of inconsistent return types
### note this no longer raises an error with dplyr >= 1.1.0
dplyr::if_else(c(TRUE, FALSE), median(1:4), median(1:5))
# Error in `dplyr::if_else()`:
# ! `false` must be a double vector, not an integer vector.

# dplyr::if_else(c(TRUE, FALSE), median_dbl(1:4), median_dbl(1:5))
# 2.5 3.0
}

\dontrun{
# stats::median raises error because of inconsistent return types
dplyr::if_else(c(TRUE, FALSE), median(1:4), median(1:5))
# Error in `dplyr::if_else()`:
# ! `false` must be a double vector, not an integer vector.

# dplyr::if_else(c(TRUE, FALSE), median_dbl(1:4), median_dbl(1:5))
# 2.5 3.0
}

}
