% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{na_if_range}
\alias{na_if_range}
\alias{na_if_range.numeric}
\alias{na_if_range.character}
\alias{na_if_range.factor}
\alias{na_if_range.labelled}
\alias{na_if_range.haven_labelled}
\title{Set NA values based on range of numbers.}
\usage{
na_if_range(x, range_min = -Inf, range_max = -1)

\method{na_if_range}{numeric}(x, range_min = -Inf, range_max = -1)

\method{na_if_range}{character}(x, range_min = -Inf, range_max = -1)

\method{na_if_range}{factor}(x, range_min = -Inf, range_max = -1)

\method{na_if_range}{labelled}(x, range_min = -Inf, range_max = -1)

\method{na_if_range}{haven_labelled}(x, range_min = -Inf, range_max = -1)
}
\arguments{
\item{x}{A numeric vector, character vector, or factor.}

\item{range_min}{The minimum value to set to \code{NA}. Defaults to \code{-Inf}.}

\item{range_max}{The maximum value to set to \code{NA}. Defaults to \code{-1}.}
}
\description{
Changes values in range \code{range_min} to \code{range_max} to \code{NA}. Works on numeric
vectors as well as for numbers in character vectors, factor labels, or in
numeric or character vectors with classes \code{labelled} or \code{haven_labelled}.
}
\note{
Previously known as \code{coerce_na_range}, which is retained as an alias for backward compatibility.
}
