% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{na_like}
\alias{na_like}
\title{Generate NA values of appropriate type}
\usage{
na_like(x, factor_as_character = TRUE, match_length = FALSE)
}
\description{
Returns compatible \code{NA} based on \code{x}. This is usually of the same type as \code{x}
(e.g., \code{NA_real_} if \code{x} is a double vector). If \code{x} is a factor, will
return \code{NA_character_} if \code{factor_as_character = TRUE} (the default) and
\code{NA_integer_} otherwise.
}
