% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{na_to_null}
\alias{na_to_null}
\alias{null_to_na}
\title{Replace NA with NULL and vice versa}
\usage{
na_to_null(x)

null_to_na(x, unlist = FALSE)
}
\description{
\code{na_to_null()} Replaces all \code{NA}s in a vector or list with \code{NULL}. Can be
useful for lists of function arguments (e.g., when using \code{purrr::pmap()}).

\code{null_to_na()} Replaces all \code{NULL}s in a list with \code{NA}s. Returns an atomic
vector if \code{unlist = TRUE} and a list otherwise.
}
