% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{next_bizday}
\alias{next_bizday}
\title{Find the next business day}
\usage{
next_bizday(
  x,
  include_today = TRUE,
  holidays = c("Chestnut", "Illinois", "federal")
)
}
\arguments{
\item{x}{A \code{Date} or coercible to \code{Date}}

\item{include_today}{If \code{x} is a business day, will return \code{x} if \code{TRUE} and
the next business day after \code{x} otherwise.}

\item{holidays}{Holiday calendar to use. Available options are "Chestnut" for
Chestnut Health System holidays, "Illinois" for State of Illinois holidays,
and "federal" for United States federal holidays.}
}
\description{
Returns the next date that isn't a weekend or holiday.
}
