% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{nth_valid}
\alias{nth_valid}
\alias{first_valid}
\alias{last_valid}
\title{Get the nth, first, or last non-\code{NA} value in a vector}
\usage{
nth_valid(x, n, order_by = NULL, default = NULL)

first_valid(x, order_by = NULL, default = NULL)

last_valid(x, order_by = NULL, default = NULL)
}
\arguments{
\item{x}{A vector.}

\item{n}{For \code{nth_valid}, the position of the value to return. Negative values start from the end of the vector. Ignored for \code{first_valid} and \code{last_valid}.}

\item{order_by}{An optional vector to use for ordering \code{x} before finding the value. Must be the same length as \code{x}.}

\item{default}{A default value to return if there are fewer than \code{n} non-\code{NA} values in \code{x}. When set to \code{NULL} (default), an error is thrown if \code{n} is beyond the number of non-\code{NA} values.}
}
\value{
\itemize{
\item \code{nth_valid}: The nth non-\code{NA} value in \code{x}.
\item \code{first_valid}: The first non-\code{NA} value in \code{x}.
\item \code{last_valid}: The last non-\code{NA} value in \code{x}.
}
}
\description{
These functions retrieve the nth, first or last non-\code{NA} value in a vector, optionally ordered by another vector. If there are fewer than n non-\code{NA} values, a default value can be returned.
}
\examples{
x <- c(NA, 2, NA, 4, 5)
nth_valid(x, n = 2)
nth_valid(x, n = 10, default = 0)
first_valid(x)
last_valid(x)

}
