% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{open_file}
\alias{open_file}
\alias{open_location}
\alias{file.open}
\alias{dir.open}
\title{Open a file or directory}
\usage{
open_file(path)

open_location(path)

file.open(path)

dir.open(path)
}
\arguments{
\item{path}{The path to the file or directory.}
}
\description{
\code{open_file()} opens the file at the specified path using the default program
for that file type. \code{open_location()} opens the file explorer at the location
of the specified file.

\code{file.open()} and \code{dir.open()} are aliases, in line with \code{base::file.create}, \code{file.exists}, \code{dir.create}, \code{dir.exists}, etc.
}
