% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{p_to_OR}
\alias{p_to_OR}
\title{Compute odds ratio from probabilities}
\usage{
p_to_OR(p1, p2)
}
\description{
Given probabilities \code{p1} and \code{p2}, returns the odds ratio (odds of \code{p2}
relative to odds of \code{p1}).
}
