% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure.R
\name{pad_vectors}
\alias{pad_vectors}
\title{Pad vectors to the same length}
\usage{
pad_vectors(...)
}
\arguments{
\item{...}{One or more vectors.}
}
\value{
A list of vectors, all having the same length as the longest input vector.
}
\description{
This function takes one or more vectors and pads them with \code{NA} values to the length of the longest vector.
}
\examples{
pad_vectors(1:3, 1:5, 1:4)

# supports list unpacking with `!!!` operator
pad_vectors(!!!list(1:3, 1:5, 1:4))


# one use case is assembling vectors of different lengths into a dataframe
# for example, to see unique column values at a glance:
unique_vals <- dplyr::starwars \%>\%
  dplyr::select(hair_color:eye_color) \%>\%
  lapply(unique)
pad_vectors(!!!unique_vals) \%>\%
  as.data.frame()

}
