% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pminmax_across}
\alias{pminmax_across}
\alias{pmax_across}
\alias{pmin_across}
\title{tidyselect-friendly parallel minima and maxima}
\usage{
pminmax_across(cols, na.rm, .fn)

pmax_across(cols, na.rm = FALSE)

pmin_across(cols, na.rm = FALSE)
}
\description{
Wrappers around \code{base::pmin()} and \code{base::pmax()} that accept
<\code{\link[=dplyr_tidy_select]{tidy-select}}> expressions.
}
\examples{
# using `base::pmax()`
mtcars \%>\%
  dplyr::mutate(
    max_val = pmax(mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb)
  )

# using `pmax_across()`
mtcars \%>\%
  dplyr::mutate(max_val = pmax_across(mpg:carb))

}
