% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{psum}
\alias{psum}
\alias{pmean}
\title{Parallel / row-wise sums and means}
\usage{
psum(..., na.rm = FALSE)

pmean(..., na.rm = FALSE)
}
\arguments{
\item{...}{one or more numeric or logical vectors of equal length or length 1.}

\item{na.rm}{logical. Should missing values (including \code{NaN}) be removed?}
}
\value{
A numeric vector.
}
\description{
Given one or more vectors, \code{psum()} and \code{pmean()} return row-wise sums or means. They are analogous to \code{\link[base:Extremes]{base::pmin()}} and \code{\link[=pmax]{pmax()}}. Like \code{pmin()} / \code{pmax()}, these functions:
\itemize{
\item accept multiple vectors via \code{...} (as opposed to \code{rowSums()} and \code{rowMeans()}, which accept a single array). This allows them to be used in data-masking contexts (e.g., inside \code{dplyr::mutate()}).
\item return \code{NA} when all values are \code{NA} and \code{na.rm = TRUE}. (as opposed to \code{rowSums()}, which returns \code{0} in this situation).
}
}
\details{
Both \code{psum()} and \code{psum_across()} can be used in data-masking contexts, but
have different capabilities and limitations. See the "Details" section of
\code{\link[=psum_across]{?psum_across}}.

Note that, unlike \code{pmin()}, \code{pmax()}, \code{rowSums()}, and \code{rowMeans()}, these
functions do not currently support objects with more than one dimension
(e.g., matrices, arrays, or data frames).
}
\examples{
psum(1:5, 6:10)
pmean(1:5, 6:10)

dat <- tibble::tribble(
  ~product,    ~price1, ~price2, ~price3,
  "Product 1", 20,      25,      22,
  "Product 2", NA,      30,      29,
  "Product 3", 15,      NA,      NA,
  "Product 4", NA,      NA,      NA
)

# contrast w `rowSums()` / `rowMeans()`:
# no need for `pick()` and different `NA` behavior
dat \%>\%
  dplyr::mutate(
    rowSums = rowSums(pick(price1, price2, price3), na.rm = TRUE),
    psum = psum(price1, price2, price3, na.rm = TRUE),
    rowMeans = rowMeans(pick(price1, price2, price3), na.rm = TRUE),
    pmean = pmean(price1, price2, price3, na.rm = TRUE)
  )

}
\seealso{
\itemize{
\item \code{\link[=psum_across]{psum_across()}} for variants that support tidyselect expressions. See the
"Details" section for relative strengths of \code{psum()} / \code{pmean()} vs.
\code{psum_across()} / \code{pmean_across()}.
\item \code{\link[base:Extremes]{base::pmin()}} for analogous \code{pmin()} and \code{pmax()} functions
\item \code{\link[=sum_if_any]{sum_if_any()}} for non-parallel sums with similar \code{NA} handling
}
}
