% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_export.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cops_go}
\alias{cops_detection_limit}
\alias{iop_go}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Cops}{\code{\link[Cops]{cops_detection_limit}}, \code{\link[Cops]{cops_go}}}

  \item{Riops}{\code{\link[Riops]{iop_go}}}
}}

