% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{reorder_dendro_by_label}
\alias{reorder_dendro_by_label}
\title{Reorder a dendrogram}
\usage{
reorder_dendro_by_label(.dendro, .order, agglo.FUN = max)
}
\description{
Reorders dendrogram leaves based on the vector of labels passed to \code{.order}.
}
