% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{reverse_key}
\alias{reverse_key}
\title{Reverse key a numeric vector}
\usage{
reverse_key(x, na.rm = FALSE, max = NULL, min = NULL)
}
\description{
Reverses a numeric vector \code{x} by subtracting from \code{min} and adding \code{max}.
Observed minimum and maximum of \code{x} are used unless otherwise specified.
}
\examples{
reverse_key(1:5)
reverse_key(3:5)
reverse_key(3:5, min = 1, max = 5)

}
