% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{safe_minmax}
\alias{safe_minmax}
\alias{safe_max}
\alias{safe_min}
\title{Maxima and minima with alternative missing value handling}
\usage{
safe_minmax(..., na.rm, .fn)

safe_max(..., na.rm = TRUE)

safe_min(..., na.rm = TRUE)
}
\description{
Returns the maximum or minimum of input values, similar to \code{base::min()} /
\code{max()}. Unlike these base functions, \code{safe_min()} and \code{safe_max()} return
\code{NA} when all values are \code{NA} and \code{na.rm = TRUE}. (\code{base::min} and \code{max()}
return \code{-Inf} or \code{Inf} in this situation). Also unlike base functions,
\code{na.rm} is \code{TRUE} by default (since this is the typical use case).
}
