% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{scale_mad}
\alias{scale_mad}
\title{Scale based on median absolute deviation}
\usage{
scale_mad(x, center = c("median", "mean", "none"), mad_constant = 1.4826)
}
\description{
Scales a vector of values based on the median absolute deviation. Values may
be centered around the median (default), mean, or not centered. Compare to
\code{base::scale()}, which uses standard deviation and centers around the mean by
default.
}
