% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{scale_vec}
\alias{scale_vec}
\title{Scaling and centering of vectors}
\usage{
scale_vec(x, center = TRUE, scale = TRUE)
}
\description{
A wrapper around \code{base::scale()} that returns a vector instead of a matrix.
}
\examples{
# using base::scale()
scale(0:4)

# using scale_vec()
scale_vec(0:4)

}
