% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{se_mean}
\alias{se_mean}
\title{Compute the standard error of the mean}
\usage{
se_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or a non-factor object coercible to numeric by \code{as.double(x)}.}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Computes the standard error of the mean for a numeric vector.
}
