% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{se_prop}
\alias{se_prop}
\title{Compute the standard error of a proportion}
\usage{
se_prop(
  x,
  na.rm = FALSE,
  min_var = 5,
  low_var_action = c("warn", "ignore", "error")
)
}
\arguments{
\item{x}{a logical or numeric vector. If numeric, it must include only 0s, 1s, and/or \code{NA}s.}

\item{na.rm}{logical. Should missing values be removed?}

\item{min_var}{numeric. Minimum variance (n * p * (1 - p) for valid normal approximation of the binomial. See Details.}

\item{low_var_action}{character. Action to take when variance is below \code{min_var}.}
}
\description{
Computes the standard error of a proportion for a logical or numeric vector.
}
\details{
Standard error of a proportion is calculated using the formula:
\deqn{SE = \sqrt{\frac{p(1 - p)}{n}}}

This formula assumes that the binomial sampling distribution underlying the
observed proportion can be approximated by a normal distribution. This
assumption is valid when the proportion variance (n\emph{p}(1 - p)) is
sufficiently large, but may not hold when variance is lower. \code{se_prop()}
therefore issues a warning when the variance is less than 5; this behavior
can be overridden using the \code{min_var} and \code{low_var_action} arguments.
}
