% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{set_compare}
\alias{set_compare}
\title{Set comparison with automatic naming}
\usage{
set_compare(x, y, autoname = TRUE)
}
\arguments{
\item{x}{A vector representing the first set.}

\item{y}{A vector representing the second set.}

\item{autoname}{Logical, whether to automatically name the returned sets based on the supplied object names. Defaults to \code{TRUE}.}
}
\value{
A named list with three elements:
\itemize{
\item Elements unique to \code{x}
\item Elements unique to \code{y}
\item Elements common to both \code{x} and \code{y}
}
}
\description{
Compares two sets (vectors), returning the elements unique to each and the elements they have in common. Optionally names the returned sets based on the supplied object names.
}
\examples{
v1 <- c(1, 2, 3, 4)
v2 <- c(3, 4, 5, 6)

set_compare(v1, v2)
set_compare(v1, v2, autoname = FALSE)

}
