% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{set_ggplot_opts}
\alias{set_ggplot_opts}
\title{Nicer default theme and palettes for \code{ggplot2}}
\usage{
set_ggplot_opts(
  base_theme = hrbrthemes::theme_ipsum_rc,
  brewer_pal_discrete = "Set1",
  ...
)
}
\description{
Changes default theme and color scales for \code{ggplot2}.
}
\details{
Theme is based on \code{hrbrthemes::theme_ipsum_rc()}, unless otherwise specified
in \code{base_theme} argument. The theme is further modified as follows:
\itemize{
\item Axis titles are centered
\item Legend title is omitted
\item Minor gridlines are omitted
\item Facet labels are placed outside axes
\item Various tweaks to text size and margins
}

Default color and fill palettes are set based on scale type:
\itemize{
\item For discrete scales, \code{RColorBrewer} palette "Set1," unless otherwise
specified in \code{brewer_pal_discrete} argument
\item For continuous and binned scales, \code{RColorBrewer} palette "Blues"
\item For ordinal scales, \code{viridisLite} palette "viridis"
}

Default font family for \code{geom_text()} and \code{geom_label()} is set to match
\code{base_theme}.
}
