% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_merge.R
\name{sql_merge}
\alias{sql_merge}
\title{sql_merge}
\usage{
sql_merge(
  projects = c("/mnt/D/Data/WISEMan", "/mnt/D/Data/CHONe", "/mnt/D/Data/IML4")
)
}
\description{
Merge multiple sqlite data bases into a new single one.
Will perform the merge table and column wise (dplyr::bind_rows). A new integer
primary key to ensure uniqueness and data tracebility. Old ID will be keep in PID column.
}
\author{
Raphael Mabit
}
