% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_collapse}
\alias{str_collapse}
\title{Collapse a character vector into a single string}
\usage{
str_collapse(..., sep = "", join = NULL)
}
\arguments{
\item{...}{one or more character vectors}

\item{sep}{string to insert between elements of each input vector}

\item{join}{optional string to combine output into a single string. join multiple collapsed vectors. If NULL (default),
returns a character vector with each input vector collapsed separately.}
}
\value{
If \code{join} is NULL, returns a character vector with length equal to the number
of vectors passed to \code{...}. If \code{join} is provided, returns a character vector
of length 1.
}
\description{
Concatenates all elements in a character vector. Essentially a column-wise
variant of dplyr::str_c(). Can accept multiple character vectors, which are
returned as separate character strings if \code{join} = NULL (the default), or
concatenated with a separator supplied to \code{join}.
}
\details{
When only a single character vector is passed to \code{...}, behavior will be
similar to \code{stringr::str_c()} using the \code{collapse} argument. But behavior
differs when multiple vectors are passed to \code{...} -- see examples. The
difference arise because \code{str_collapse()} first collapses each vector, then
optionally joins the resulting vectors, whereas \code{stringr::str_c()} first
joins across vectors before collapsing the resulting vector.
}
\examples{
# with just a single character vector, behavior is similar to `stringr::str_c()`
# with the `collapse` argument
abc <- c("a", "b", "c")
str_collapse(abc, sep = "-")
stringr::str_c(abc, collapse = "-")

# but behavior differs when multiple vectors are passed
def <- c("d", "e", "f")
str_collapse(abc, def, sep = "-")
stringr::str_c(abc, def, collapse = "-")

str_collapse(abc, def, sep = "-", join = " | ")
stringr::str_c(abc, def, collapse = "-", sep = " | ")
stringr::str_c(abc, def, sep = "-", collapse = " | ")

# can accept vectors of different lengths
lmnop <- c("l", "m", "n", "o", "p")
str_collapse(abc, def, lmnop, join = ", ")

}
