% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_prefix}
\alias{str_prefix}
\alias{str_suffix}
\title{Find common prefixes or suffixes}
\usage{
str_prefix(string, na.rm = FALSE)

str_suffix(string, na.rm = FALSE)
}
\description{
Returns substring beginnings or endings common to all elements in vector.
}
\examples{
test_words <- c("antidote", "antimony", "antimatter", "antisense")
str_prefix(test_words)

wdays <- c(
  "Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"
)
str_suffix(wdays)

}
