% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_starts_any}
\alias{str_starts_any}
\title{Detect the presence of any pattern at the beginning of a string}
\usage{
str_starts_any(string, patterns, ignore_case = FALSE, negate = FALSE)
}
\description{
Returns TRUE if any element of \code{patterns} is present in \code{string}.
}
