% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summaries.R
\name{summary_table}
\alias{summary_table}
\title{Custom summary table}
\usage{
summary_table(
  .data,
  ...,
  .vars = where(is.numeric),
  na.rm = FALSE,
  .rows_group_by = NULL,
  .cols_group_by = NULL,
  .cols_group_glue = NULL,
  .cols_group_order = c("by_group", "by_function"),
  .var_col_name = "Variable",
  .cols_group_opts = list()
)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a tibble).}

\item{...}{Functions to apply to each variable specified in \code{.vars}, as
function names (e.g., \code{mean}) or anonymous functions (e.g.,
\verb{\\(x) mean(x, na.rm = TRUE)}). Each function passed to \code{...} will yield one
column in the output table (or one column per group if \code{.cols_group_by} is
specified). Output column names can optionally be specified. E.g.,
\verb{m = mean, sd, sem = \\(x) sd(x) / sqrt(n())} yields output columns named \code{m},
\code{sd}, and \code{sem}.}

\item{.vars}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns in \code{.data} to
summarize. Each column passed to \code{.vars} will yield one row in the output
table (or one row per group if \code{.rows_group_by} is specified).}

\item{na.rm}{A logical value passed to any functions in \code{...} that take an
\code{na.rm} argument.}

\item{.rows_group_by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Grouping
variable(s) for output rows.}

\item{.cols_group_by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Grouping
variable(s) for output columns.}

\item{.cols_group_glue}{A glue specification controlling column names when
\code{.cols_group_by} is specified. See examples.}

\item{.cols_group_order}{When \code{.cols_group_by} is specified, how should the resulting columns be ordered?
\itemize{
\item "by_group" keeps columns from each group together, resulting in columns ordered like: fn1_group1, fn2_group1, fn1_group2, fn2_group2. This is the default.
\item "by_function" keeps columns from each function together, resulting in columns ordered like: fn1_group1, fn1_group2, fn2_group1, fn2_group2.
}}

\item{.var_col_name}{The name of the output column containing variable names
passed to \code{.vars}. This column will be dropped if \code{.var_col_name} is \code{NULL} and only one variable is passed to \code{.vars}.}

\item{.cols_group_opts}{DEPRECATED. A list of additional arguments passed to
\code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}} if \code{.cols_group_by} is specified. This argument will
be removed in a future release.}
}
\description{
Generates a summary table, with a row for each variable passed to \code{.vars} and
a column for each function passed to \code{...}.
}
\examples{
# example data
mtcars2 <- mtcars \%>\%
  dplyr::mutate(
    Transmission = dplyr::recode(am, `0` = "auto", `1` = "manual")
  )

# simple summary table. note specification of column and row names
# for "n", "m", and "weight".
mtcars2 \%>\%
  summary_table(
    n = n_valid, m = mean, sd,
    .vars = c(mpg, hp, weight = wt)
  )

# with column and row groupings
mtcars2 \%>\%
  summary_table(
    n = n_valid, m = mean, sd,
    .vars = c(mpg, hp, weight = wt),
    .cols_group_by = cyl,
    .rows_group_by = Transmission
  )

# `.var_col_name = NULL` will drop the variable name column if only one
# variable is included in `.vars`
mtcars2 \%>\%
 summary_table(
   n = n_valid, m = mean, sd,
   .vars = mpg,
   .cols_group_by = cyl,
   .rows_group_by = Transmission,
   .var_col_name = NULL
 )

# customize column names with a glue specification:
mtcars2 \%>\%
  summary_table(
    M = mean, SD = sd,
    .vars = c(mpg, hp, weight = wt),
    .cols_group_by = c(Transmission, cyl),
    .cols_group_glue = "{cyl} cyl {Transmission}: {.value}"
  )

}
