% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{suppress_if}
\alias{suppress_if}
\alias{suppress_warnings_if}
\alias{suppress_messages_if}
\title{Conditionally suppress warnings or messages}
\usage{
suppress_warnings_if(
  expr,
  msg_contains = "",
  fixed = TRUE,
  perl = !fixed,
  ignore.case = FALSE,
  negate = FALSE,
  classes = "warning"
)

suppress_messages_if(
  expr,
  msg_contains = "",
  fixed = TRUE,
  perl = !fixed,
  ignore.case = FALSE,
  negate = FALSE,
  classes = "message"
)
}
\arguments{
\item{expr}{The expression to evaluate.}

\item{msg_contains}{A string to match against the message. Defaults to an empty string (matches all messages).}

\item{fixed}{A logical indicating whether \code{msg_contains} should be matched as a fixed string..}

\item{perl}{A logical indicating whether Perl-compatible regular expressions should be used for \code{msg_contains}.}

\item{ignore.case}{A logical indicating whether the case of \code{msg_contains} should be ignored.}

\item{negate}{A logical indicating whether the message matching should be negated (e.g., suppress messages that do \emph{not} match \code{msg_contains}).}

\item{classes}{A character vector of warning or message classes to suppress.}
}
\value{
The result of evaluating \code{expr}, with specified warnings or messages suppressed.
}
\description{
Evaluates an expression while selectively suppressing warnings or messages based on their content.
}
\examples{
# Suppress warnings containing specific text
suppress_warnings_if(warning("This is a warning"), "This")

# Suppress messages unless they contain specific text
suppress_messages_if(
  message("13 files processed"),
  "\\\\d{2, }",
  fixed = FALSE,
  negate = TRUE
)

}
