% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_filter.R
\name{swap}
\alias{swap}
\alias{swap_if}
\title{Swap column values, optionally based on condition}
\usage{
swap(.data, x, y)

swap_if(.data, cond, x, y, missing = c("NA", "keep", "swap"))
}
\arguments{
\item{x}{column to swap with \code{y}.}

\item{y}{column to swap with \code{x}.}

\item{cond}{a logical vector.}

\item{missing}{what to do if \code{cond} is \code{NA}.
\itemize{
\item \code{"NA"} replaces both \code{x} and \code{y} with \code{NA}.
\item \code{"keep"} keeps values in their original columns (as though \code{cond} were \code{FALSE}).
\item \code{"swap"} swaps values between \code{x} and \code{y} (as though \code{cond} were \code{TRUE}).
}}
}
\description{
\code{swap()} swaps values between two columns. \code{swap_if()} swaps values only in rows where a condition is met. \code{x} and \code{y} will be cast to a common type using \code{vctrs::vec_cast_common()}.
}
\references{
Based on Hadley's code in a \href{https://github.com/tidyverse/dplyr/issues/2149#issuecomment-258916706}{this dplyr GitHub issue}.
}
