% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{swap_if}
\alias{swap_if}
\title{Swap column values based on condition}
\usage{
swap_if(.data, cond, x, y)
}
\description{
Swaps values between two columns if condition is met. Based on Hadley's code
at: https://github.com/tidyverse/dplyr/issues/2149#issuecomment-258916706
}
