% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{syms_to_chr}
\alias{syms_to_chr}
\title{Print symbols as a character vector}
\usage{
syms_to_chr(..., width = 80, indent = 0, indent_first = indent)
}
\arguments{
\item{...}{symbols.}

\item{width}{maximum width of each line of output.}

\item{indent}{number of spaces to indent.}

\item{indent_first}{number of spaces to indent the first line.}
}
\description{
A helper for converting a list of symbols to a character vector. Primarily
intended to help convert old code given updates to \code{summary_report()}.
}
\examples{
syms_to_chr(
  various, very, video, view, village, visit, vote, wage, wait, walk, wall,
  want, war, warm, wash, waste, watch, water, way, we, wear, wednesday, wee,
  week, weigh, welcome, well, west, what, when, where, whether, which, white,
  who, whole, why, wide, wife, will, win, wind, window, wish, with, within,
  without, woman, wonder, wood, word, work, world, worry, worse, worth, would,
  write, wrong, year, yes, yesterday, yet, you, young
)

}
