% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure.R
\name{t_tibble}
\alias{t_tibble}
\title{Transpose a tibble}
\usage{
t_tibble(x, names_to = "Variable", names_from = NULL)
}
\arguments{
\item{x}{A dataframe or tibble.}

\item{names_to}{Name for column in the transposed tibble containing column
names from \code{x}}

\item{names_from}{Column in \code{x} used for column names of transposed tibble.
If not specified and \code{x} has rownames, these will be used; otherwise columns
will be named \code{...1}, \code{...2}, etc.}
}
\description{
Given a tibble or data.frame \code{x}, returns the transpose of \code{x}. Similar to
\code{base::t()}, but accepts and returns tibbles, and includes options for row
column name handling.
}
