% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{territory}
\alias{territory}
\alias{state.terr.name}
\alias{state.terr.abb}
\alias{state.terr.data}
\title{US state and territory data}
\format{
An object of class \code{character} of length 56.

An object of class \code{character} of length 56.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 56 rows and 4 columns.
}
\usage{
state.terr.name

state.terr.abb

state.terr.data
}
\description{
\code{state.terr.name} and \code{state.terr.abb} expand the built-in \code{state.name} and
\code{state.abb} vectors by adding US territories and the District of Columbia.
This includes:
\itemize{
\item American Samoa
\item District of Columbia
\item Guam
\item Northern Mariana Islands
\item Puerto Rico
\item Virgin Islands
}

\code{state.terr.data} includes names, abbreviations, and FIPS codes for US states
and territories.
}
\seealso{
\code{\link[datasets]{state.name}}
}
\keyword{datasets}
