% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{unpack-assign}
\alias{unpack-assign}
\alias{\%<-\%}
\alias{\%->\%}
\title{Unpack and assign}
\usage{
lhs \%<-\% rhs

lhs \%->\% rhs
}
\description{
Assigns a vector of values to an equal-length vector of names.
}
\details{
The \verb{\%<-\%} operator and its left-to-right variant (\verb{\%->\%}) allow parallel
assignment, similar to (e.g.) Python:\preformatted{# python
var1, var2, var3 = [2, 4, 6]

# R
c(var1, var2, var3) \%<-\% c(2, 4, 6)
}

See the \code{zeallot} package for a
similar operator with more advanced functionality.
}
\examples{
c(cyl4, cyl6, cyl8) \%<-\% split(mtcars, mtcars$cyl)
cyl4

split(mtcars, mtcars$gear) \%->\% c(G3, G4, G5)
G5

c(model_hp, model_size, model_trans) \%<-\% purrr::map(
  c("hp", "disp + wt", "gear * am"),
  ~ lm(paste("mpg ~", .x), data = mtcars)
)
summary(model_size)
summary(model_trans)

# cleanup
rm(cyl4, cyl6, cyl8, G3, G4, G5, model_hp, model_size, model_trans)

}
