% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize extreme values}
\usage{
winsorize(
  x,
  max_dev = 3,
  method = c("sd", "mad"),
  mad.center = c("median", "mean")
)
}
\description{
Sets all values more than \code{max_dev} deviations from center to be \code{max_dev}
deviations from center. Deviations defined as standard deviation (the
default) or mean absolute deviation (if \code{method = "mad"}). Center defined as
mean for \code{method = "sd"} and median for \code{method = "mad"}, unless otherwise
specified in \code{center} argument.
}
