% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{zap_everything}
\alias{zap_everything}
\title{Strip special attributes from SPSS dataset}
\usage{
zap_everything(.data, ..., .as_factor = TRUE)
}
\description{
Removes all special attributes from data read from SPSS, optionally
converting labelled vectors to factors if \code{.as_factor} = TRUE (the default).
}
