% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{asterisks}
\alias{asterisks}
\title{Return asterisks corresponding to p-values}
\usage{
asterisks(
  p,
  trends = TRUE,
  levels = c(0.1, 0.05, 0.01, 0.001),
  marks = c(sig = "*", trend = "+", ns = NA_character_),
  include_key = FALSE
)
}
\arguments{
\item{p}{a numeric vector of p-values.}

\item{trends}{logical. Should trends (e.g., .05 < p < .10) be marked?}

\item{levels}{a numeric vector demarcating ranges of p values that should receive unique significance marks.}

\item{marks}{a named character vector specifying marks for significance, trend, and non-significance.}

\item{include_key}{logical. should a key for significance marks be included as an attribute?}
}
\value{
A character vector of asterisks corresponding to the p-values. If \code{include_key = TRUE}, the vector will have a 'key' attribute indicating the significance levels.
}
\description{
Returns asterisks indicating significance levels for a vector of p values.
}
\examples{
p <- c(0.5, 0.09, 0.03, 0.008, 0.0003)

tibble::tibble(p, sig = asterisks(p))

tibble::tibble(p, sig = asterisks(p, trends = FALSE))

tibble::tibble(p, sig = asterisks(p, trends = FALSE, marks = c(ns = "ns")))

asterisks(p, include_key = TRUE)

}
