% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{bizday}
\alias{bizday}
\alias{nth_bizday}
\alias{next_bizday}
\title{Find the nth or next business day}
\usage{
nth_bizday(
  x,
  n,
  include_today = FALSE,
  holidays = c("Chestnut", "Illinois", "federal")
)

next_bizday(
  x,
  include_today = FALSE,
  holidays = c("Chestnut", "Illinois", "federal")
)
}
\arguments{
\item{x}{a date or vector of dates.}

\item{n}{integer indicating ow many business days forward to find.}

\item{include_today}{logical indicating whether \code{x} be counted as one day (assuming it's a business day)?}

\item{holidays}{character indicating which set of holidays to use.}
}
\value{
\itemize{
\item \code{nth_bizday} returns the nth business day from the provided date(s).
\item \code{next_bizday} returns the next business day from the provided date(s).
}
}
\description{
\code{nth_bizday()} returns the nth business day from a given date, based on CHS, Illinois, or federal holidays. \code{next_bizday()} is a wrapper for
}
\examples{
next_bizday(as.Date("2024-07-02"))
nth_bizday(as.Date("2024-07-02"), 5)
nth_bizday(as.Date("2024-07-02"), 5, include_today = TRUE)

}
