% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{coerce_na_range}
\alias{coerce_na_range}
\title{Set NA values based on numbers stored as strings - DEPRECATED}
\usage{
coerce_na_range(x, range_min = -Inf, range_max = -1)
}
\description{
Deprecated in lighthouse 0.7.2 favor of \code{\link[=na_if_range]{na_if_range()}}, which
this now calls.

Changes values coercible to numeric in range \code{range_min}:\code{range_max} to \code{NA}.
Useful for imported SPSS files.
}
