% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{count_multiple}
\alias{count_multiple}
\title{Count observations for multiple variables}
\usage{
count_multiple(
  .data,
  ...,
  .pct = TRUE,
  sort = FALSE,
  name = NULL,
  na.rm = FALSE,
  .drop = TRUE
)
}
\description{
Returns a table with frequencies for each column passed to \code{...}.
}
