% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{count_pct}
\alias{count_pct}
\title{Count observations with percentage}
\usage{
count_pct(.data, ..., na.rm = FALSE, multiply = FALSE)
}
\description{
Adds column with percent of total to results of \code{\link[=count]{dplyr::count()}}.
}
