% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{count_unique}
\alias{count_unique}
\title{Count unique values in data frame columns}
\usage{
count_unique(.data, ..., name = "n_unique", na.rm = FALSE)
}
\arguments{
\item{.data}{a data frame.}

\item{...}{columns to count unique values across.}

\item{name}{name to give the unique count column.}

\item{na.rm}{exclude \code{NA}s from counts?}
}
\description{
This variant of \code{\link[dplyr:count]{dplyr::count()}} returns the number of unique values across a set of columns in a data frame.
}
\examples{
mtcars \%>\%
  count_unique(cyl, gear)

mtcars \%>\%
  count_unique(cyl, gear, carb, name = "unique_combos")

}
