% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{count_with_total}
\alias{count_with_total}
\title{Count observations with total}
\usage{
count_with_total(
  .data,
  ...,
  totals_for = NULL,
  label = "Total",
  first_row = FALSE,
  .drop = TRUE
)
}
\description{
Adds row with column totals to results of \code{\link[=count]{dplyr::count()}}. Can take
multiple variables, but totals only by first variable given.
}
