% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{crosstab}
\alias{crosstab}
\title{Cross-tabulate observations}
\usage{
crosstab(.data, ..., .drop = TRUE)
}
\description{
Builds a contingency table similar to \code{base::table()}. Unlike
\code{base::table()}, \code{crosstab()} is pipe-friendly, and outputs an ordinary
tibble / data.frame -- e.g., it will retain its structure if exported to csv.

Currently supports only two variables.
}
