% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cumsum_desc}
\alias{cumsum_desc}
\title{Descending cumulative sum}
\usage{
cumsum_desc(x)
}
\description{
Returns a cumulative sum beginning with the last element of \code{x}.
}
\examples{
diamonds \%>\%
  count(cut) \%>\%
  mutate(
    or_worse = cumsum(n),
    or_better = cumsum_desc(n)
  )

}
