% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_filter.R
\name{fct_case_when}
\alias{fct_case_when}
\title{Results of case_when() as factor.}
\usage{
fct_case_when(..., .ordered = FALSE)
}
\description{
Wrapper for \code{\link[=case_when]{dplyr::case_when()}}, but result is a factor with
levels in same order passed to \code{...}. Returns an ordered factor if \code{.ordered}
is \code{TRUE}.
}
