% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{fiscal_year}
\alias{fiscal_year}
\alias{ffy}
\alias{sfy_il}
\title{Determine fiscal year from date}
\usage{
fiscal_year(x, type = c("year", "date_first", "date_last"), fiscal_start = 1)

ffy(x, type = c("year", "date_first", "date_last"))

sfy_il(x, type = c("year", "date_first", "date_last"))
}
\arguments{
\item{x}{A date or date-time object.}

\item{type}{What to return: the fiscal year ("year"), first day of the fiscal year ("date_first"), or last day of the fiscal year ("date_last").}

\item{fiscal_start}{For \code{fiscal_year}, the month the fiscal year starts (default is 1 for January).}
}
\value{
An integer representing the fiscal year or a Date representing the start or end of the fiscal year, depending on \code{type}.
}
\description{
Given a date, returns the corresponding fiscal year, start date, or end date. The \code{fiscal_year} function allows specifying the fiscal year start month, while \code{ffy} and \code{sfy_il} are convenience wrappers:
\itemize{
\item \code{ffy}: Federal fiscal year (starts in October)
\item \code{sfy_il}: Illinois state fiscal year (starts in July)
}
}
\examples{
dt <- as.Date("2023-08-15")
fiscal_year(dt)
fiscal_year(dt, fiscal_start = 7)
fiscal_year(dt, type = "date_first", fiscal_start = 7)
fiscal_year(dt, type = "date_last", fiscal_start = 7)
ffy(dt)
sfy_il(dt)

}
