% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_iop_db.R
\name{generate_iop_db}
\alias{generate_iop_db}
\title{generate instrument databases in L3}
\usage{
generate_iop_db(project, mission = "XXX", boat = c(""))
}
\description{
Loop over all L2 Station QC above 0 (in iop_processing_log) to load and extract
all non-empty list in IOP.fitted.down.RData file.
Create in L3 one csv file by instrument and a global html report for all processed variables.
}
\author{
Raphael Mabit
}
