% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_col_types}
\alias{get_col_types}
\title{Summarize a dataframe's column types - DEPRECATED}
\usage{
get_col_types(.data)
}
\description{
Returns class and type for each column in \code{.data}. Deprecated in favor of
\code{cols_info()}, which provides more information and features and is more
stable.
}
