% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{glue_chr}
\alias{glue_chr}
\title{Format and interpolate a string as character vector}
\usage{
glue_chr(...)
}
\arguments{
\item{...}{[\code{expressions}]\cr Unnamed arguments are taken to be expression
string(s) to format. Multiple inputs are concatenated together before formatting.
Named arguments are taken to be temporary variables available for substitution.

\if{html}{\out{<div class="sourceCode">}}\preformatted{For `glue_data()`, elements in `...` override the values in `.x`.
}\if{html}{\out{</div>}}}
}
\description{
A wrapper around \code{glue::glue()} which returns a character vector rather than
a \code{"glue"} object.
}
