% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{in_excel}
\alias{in_excel}
\title{Open dataframe in Excel}
\usage{
in_excel(df, name = NULL, na = "")
}
\arguments{
\item{df}{a data frame or data frame extension (e.g., a tibble).}

\item{name}{a string: the name to use for the .csv file. If \code{NULL}, a random name will be generated.}

\item{na}{string to use for missing values in the .csv file.}
}
\description{
Saves dataframe as .csv in R temp directory, then opens in Excel. The .csv
will have a randomly-generated name unless otherwise specified in \code{name}.
}
