% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{in_excel}
\alias{in_excel}
\title{Open dataframe in Excel}
\usage{
in_excel(df, name, na = "")
}
\arguments{
\item{df}{The dataframe to open in Excel.}

\item{name}{(Optional) The name to use for the .csv file. If not provided, a random name will be generated.}

\item{na}{(Optional) The string to use for missing values in the .csv file. Defaults to an empty string.}
}
\description{
Saves dataframe as .csv in R temp directory, then opens in Excel. The .csv
will have a randomly-generated name unless otherwise specified in \code{name}.
}
\examples{
\\dontrun{
  iris |> in_excel()
  mtcars |> in_excel(\"mtcars_data\")
}
}
