% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_duplicate}
\alias{is_duplicate}
\title{Identify duplicates within a vector or set of vectors.}
\usage{
is_duplicate(..., nmax = 1, incomparables = FALSE)
}
\description{
Identify duplicates within a vector or set of vectors.
}
\examples{
x <- c(1, 2, 2, 3, 3, 3)
y <- c(1, 1, 2, 1, 2, 2)
is_duplicate(x)
is_duplicate(x, y)

}
