% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{is_duplicate}
\alias{is_duplicate}
\title{Identify duplicates within a vector or vectors}
\usage{
is_duplicate(..., nmax = 1, incomparables = FALSE)
}
\arguments{
\item{...}{one or more vectors of equal length.}

\item{nmax}{maximum number of times a value can appear before being considered a duplicate.}

\item{incomparables}{should missing values (including \code{NaN}) be considered duplicates?}
}
\value{
a logical vector.
}
\description{
This function checks for duplicated values within a vector or a set of vectors.
}
\examples{
x <- c(1, 2, 2, 3, 3, 3)
y <- c(1, 1, 2, 1, 2, 2)
is_duplicate(x)
is_duplicate(x, nmax = 2)
is_duplicate(x, y)

z <- c(1, NA, NA)
is_duplicate(z)
is_duplicate(z, incomparables = TRUE)

}
