% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{is_spss}
\alias{is_spss}
\title{Test whether a data frame contains SPSS variable or value labels}
\usage{
is_spss(.data)
}
\arguments{
\item{.data}{A data frame.}
}
\value{
\code{TRUE} if the data frame contains any SPSS labels, \code{FALSE} otherwise.
}
\description{
Checks if a data frame contains any SPSS / haven variable labels, value labels, or format attributes.
}
