% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{median_dbl}
\alias{median_dbl}
\title{Median value as double}
\usage{
median_dbl(x, na.rm = FALSE, ...)
}
\description{
Returns median of \code{x} as a double vector. Alternative to \code{stats::median()}
when a consistent return value is needed.
}
\examples{
\dontrun{
# stats::median raises error because of inconsistent return types
dplyr::if_else(c(TRUE, FALSE), median(1:4), median(1:5))
# Error in `dplyr::if_else()`:
# ! `false` must be a double vector, not an integer vector.

# dplyr::if_else(c(TRUE, FALSE), median_dbl(1:4), median_dbl(1:5))
# 2.5 3.0
}

}
