% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{n_valid}
\alias{n_valid}
\alias{pct_valid}
\alias{n_pct_valid}
\title{Count non-missing cases}
\usage{
n_valid(x, out = c("n", "pct", "n_pct"), ...)

pct_valid(x, ...)

n_pct_valid(x, ...)
}
\description{
\code{n_valid()} returns the number of vector elements that are not \code{NA}. It
returns the percentage of non-\code{NA} values if \code{out} = \code{"pct"}, and a tibble
containing number and percentage if \code{out} = \code{"n_pct"}.

\code{pct_valid()} is a wrapper around \code{n_valid(out = "pct")}.

\code{n_pct_valid()} is a wrapper around \code{n_pct_valid(out = "n_pct")}.
}
