% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{nth_valid}
\alias{nth_valid}
\alias{first_valid}
\alias{last_valid}
\title{Get the nth, first, or last non-\code{NA} value in a vector}
\usage{
nth_valid(x, n, default = NA)

first_valid(x, default = NA)

last_valid(x, default = NA)
}
\arguments{
\item{x}{A vector.}

\item{n}{integer. Position of non-\code{NA} value to return. Negative values start from the end of the vector.}

\item{default}{a default value to use if there are fewer than \code{n} non-\code{NA} values in \code{x}. Will be cast to the type of \code{x}.}
}
\value{
\itemize{
\item \code{nth_valid}: The nth non-\code{NA} value in \code{x}.
\item \code{first_valid}: The first non-\code{NA} value in \code{x}.
\item \code{last_valid}: The last non-\code{NA} value in \code{x}.
}
}
\description{
These functions retrieve the nth, first or last non-\code{NA} value in a vector. If there are fewer than \code{n} non-\code{NA} values, a default value can be returned.
}
\examples{
x <- c(NA, 7, NA, 5, 4, NA, 2, NA)

first_valid(x)
last_valid(x)

nth_valid(x, 2)
nth_valid(x, -2)

nth_valid(x, 6)
nth_valid(x, 6, default = -Inf)

}
\seealso{
{
\code{\link[dplyr:nth]{dplyr::nth()}}
}
}
