% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{open_file}
\alias{open_file}
\alias{open_location}
\alias{file.open}
\alias{dir.open}
\title{Open a file or directory}
\usage{
open_file(path)

open_location(path)

file.open(path)

dir.open(path)
}
\arguments{
\item{path}{The path to the file or directory.}
}
\description{
Functions to open a file with its default program or open a file's location in the file explorer.
}
\details{
\code{file.open()} and \code{dir.open()} are aliases, in line with \code{base::file.create}, \code{file.exists}, \code{dir.create}, \code{dir.exists}, etc.
}
\section{Functions}{

\itemize{
\item \code{open_file()}: Opens the file at the specified path using the default program for that file type. Automatically handles paths with special characters.
\item \code{open_location()}: Opens the file explorer at the location of the specified file, with the file selected if possible. Automatically handles paths with special characters.
}
}

\examples{
\dontrun{
# Open a file
open_file("path/to/file.txt")

# Open a file's location
open_location("path/to/file.txt")
}

}
