% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{p_to_OR}
\alias{p_to_OR}
\alias{OR_to_p2}
\alias{OR_to_p1}
\title{Convert between probabilities and odds ratios}
\usage{
p_to_OR(p1, p2)

OR_to_p2(p1, OR)

OR_to_p1(p2, OR)
}
\arguments{
\item{p1}{The first probability.}

\item{p2}{The second probability.}

\item{OR}{The odds ratio.}
}
\value{
\itemize{
\item \code{p_to_OR}: The odds ratio corresponding to the given probabilities.
\item \code{OR_to_p2}: The second probability corresponding to the given first probability and odds ratio.
\item \code{OR_to_p1}: The first probability corresponding to the given second probability and odds ratio.
}
}
\description{
These functions convert between probabilities and odds ratios. \code{p_to_OR} calculates the odds ratio given two probabilities, while \code{OR_to_p2} and \code{OR_to_p1} calculate the second and first probability, respectively, given the other probability and the odds ratio.
}
\examples{
p_to_OR(0.4, 0.6)
OR_to_p2(0.4, 2.25)
OR_to_p1(0.6, 2.25)

}
