% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{rbool}
\alias{rbool}
\title{Generate random logicals}
\usage{
rbool(n, prob = 0.5)
}
\description{
Returns a vector of random logicals of length \code{n}, drawn from binomial
distribution with trial probability \code{prob} (default = .5).
}
