% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{row_sums_across}
\alias{row_sums_across}
\title{Row sums for selected columns with \code{NA} handling}
\usage{
row_sums_across(cols, na.rm = FALSE)
}
\arguments{
\item{cols}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> columns to sum across.}

\item{na.rm}{Should missing values (including \code{NaN}) be removed?}
}
\description{
This function calculates row sums for selected columns using tidyselect expressions. Unlike \code{rowSums}, it returns \code{NA} rather than \code{0} when \code{na.rm = TRUE} and all selected columns are \code{NA}.
}
\examples{
df <- tibble::tibble(
  x = c(1, 2, NA, NA),
  y = c(5, NA, 7, NA),
  z = c(9, 10, 11, NA)
)

df \%>\%
  dplyr::mutate(
    row_sums = row_sums_across(x:z),
    row_sums_na.rm = row_sums_across(x:z, na.rm = TRUE)
  )

}
