% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{row_sums_spss}
\alias{row_sums_spss}
\title{Replicate SPSS SUM() function}
\usage{
row_sums_spss(...)
}
\description{
Sums across columns a la SPSS: \code{NA}s are counted as 0s, but if ALL variables
are \code{NA}, then result is \code{NA}.
}
