% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{safe_minmax}
\alias{safe_minmax}
\alias{safe_max}
\alias{safe_min}
\title{Maxima and minima with alternative missing value handling - DEPRECATED}
\usage{
safe_max(..., na.rm = TRUE)

safe_min(..., na.rm = TRUE)
}
\description{
Deprecated in lighthouse 0.7.0 favor of \code{\link[=max_if_any]{max_if_any()}} and \code{\link[=min_if_any]{min_if_any()}}, which
these now call.
}
