% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{se}
\alias{se}
\title{Compute the standard error}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an R object but not a \code{factor} coercible to
numeric by \code{as.double(x)}.}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Computes the standard error of the values in x.
}
