% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_collapse}
\alias{str_collapse}
\title{Collapse a character vector into a single string}
\usage{
str_collapse(..., sep = "", join = NULL)
}
\description{
Concatenates all elements in a character vector. Essentially a column-wise
variant of dplyr::str_c(). Can accept multiple character vectors, which are
returned as separate character strings if \code{join} = NULL (the default), or
concatenated with a separator supplied to \code{join}.
}
