% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_detect_any}
\alias{str_detect_any}
\title{Detect the presence of any pattern in a string}
\usage{
str_detect_any(
  string,
  patterns,
  whole_word = FALSE,
  ignore_case = FALSE,
  negate = FALSE
)
}
\description{
Returns TRUE if any element of \code{patterns} is present in \code{string} (for
\code{str_detect_any()}) or at the start or end of \code{string} (for
\code{str_starts_any()} and \code{str_ends_any()}). Matches whole words if \code{whole_word}
= TRUE (FALSE by default). Case sensitive unless \code{ignore_case} = TRUE (FALSE
by default).
}
