% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_detect_any}
\alias{str_detect_any}
\alias{str_starts_any}
\alias{str_ends_any}
\title{Detect the presence of any pattern in a string}
\usage{
str_detect_any(
  string,
  patterns,
  whole_word = FALSE,
  ignore_case = FALSE,
  negate = FALSE
)

str_starts_any(
  string,
  patterns,
  whole_word = FALSE,
  ignore_case = FALSE,
  negate = FALSE
)

str_ends_any(
  string,
  patterns,
  whole_word = FALSE,
  ignore_case = FALSE,
  negate = FALSE
)
}
\arguments{
\item{string}{a character vector or something coercible to one.}

\item{patterns}{a character vector containing regular expressions to look for.}

\item{whole_word}{logical. Match only whole words in \code{string} (as defined by \code{"\\\\b"} word boundary)?}

\item{ignore_case}{logical. Ignore case when matching \code{patterns} to \code{string}?}

\item{negate}{Logical. if \code{TRUE}, inverts the resulting boolean vector.}
}
\description{
\code{str_detect_any()} returns TRUE if any element of \code{patterns} is present in
\code{string}. \code{str_starts_any()} and \code{str_ends_any()} match only at the beginning
or end of strings.
}
\seealso{
{
\code{\link[stringr:str_detect]{stringr::str_detect()}}
\code{\link[stringr:str_starts]{stringr::str_starts()}}
}
}
