% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{strftime_no_lead}
\alias{strftime_no_lead}
\title{Format date-time to string without leading zeros}
\usage{
strftime_no_lead(x, format = "\%m/\%d/\%Y", tz = "", usetz = FALSE, ...)
}
\arguments{
\item{x}{A date-time object.}

\item{format}{A character string giving a date-time format as used by \code{strftime()}.}

\item{tz}{A character string specifying the time zone to be used.}

\item{usetz}{A logical value indicating whether the time zone abbreviation should be appended to the output.}

\item{...}{Further arguments to be passed to \code{format.POSIXlt()}.}
}
\value{
A character vector representing the date-time without leading zeros.
}
\description{
Converts a date-time object to a character string without leading zeros in numeric components. Wraps \code{format.POSIXlt()}, removing leading zeros using regex substitution.
}
\examples{
dt <- as.POSIXct("2023-06-05 01:02:03")

# With leading zeros
format(dt, "\%m/\%d/\%Y \%H:\%M:\%S")

# Without leading zeros
strftime_no_lead(dt, "\%m/\%d/\%Y \%H:\%M:\%S")

}
